Refactor error types to be cleaner
This changes the way that the error types are defined such that the
error names and formatting are handled by the enum, with only the data
being split out into separate structs. This allows reuse of the internal
data and simplifies creating the error values. Constructors for the new
error types are also provided to further streamline their creation.