Added maximum stack size and checks
This adds a new `pushstack!(...)` macro for pushing which first checks
that the number of elements requested will fit in the stack, then pushes
each of them. This uses another new macro `checkstack!(...)` which is
intended for use with other opcodes which need to do more complex stack
operations as well.

Only operations which increase the stack size need check the stack size;
other operations such as `add` and `add.c` either decrease the stack
size or keep the stack size the same and as such don't need this check
before directly using the stack.

A more complicated macro which allowed for arbitrary numbers of
arguments to be pushed onto the stack was considered and discarded, as
this macro is mostly for use with a fixed number of values; it's simple
enough to add new cases, or use the generic case with

    `pushstack!(@n self, code[n], "opcode", 5, v1, v2, v3, v4, v5);`